requirements.txt
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
google-api-core==2.10.1
google-auth==2.11.0
google-cloud-compute==1.5.2
googleapis-common-protos==1.56.4
grpcio==1.49.0
grpcio-status==1.49.0
idna==3.4
proto-plus==1.22.1
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.1
rsa==4.9
six==1.16.0
urllib3==1.26.12
